Reorganize OpenLIFULib (#61)
As we added more and more parameter node serializers for the
work of Issue #61, OpenLIFULib's __init__ got way too busy.
So I have moved the more mundane parameter node related stuff
to parameter_node_utils.py. As a consequence a few other items
needed to be moved and reorganized.
