+++ OpenLIFULib/CMakeLists.txt
@@ -4,7 +4,10 @@ set(MODULE_NAME OpenLIFULib)
 #-----------------------------------------------------------------------------
 set(MODULE_PYTHON_SCRIPTS
   OpenLIFULib/__init__.py
-  )
+  OpenLIFULib/busycursor.py
+  OpenLIFULib/lazyimport.py
+  OpenLIFULib/parameter_node_utils.py
+)
 
 set(MODULE_PYTHON_RESOURCES
  OpenLIFULib/Resources/python-requirements.txt
