+++ OpenLIFULib/OpenLIFULib/__init__.py
@@ -1,11 +1,10 @@
 import qt
 import vtk
-from typing import Any, List, Tuple, Sequence, TYPE_CHECKING, get_origin, get_args, NamedTuple
+from typing import Any, List, Sequence, TYPE_CHECKING, NamedTuple
 from scipy.ndimage import affine_transform
 from vtk.util import numpy_support
 import numpy as np
 from numpy.typing import NDArray
-from pathlib import Path
 import slicer
 from slicer import (
     vtkMRMLModelNode,
@@ -13,93 +12,39 @@ from slicer import (
     vtkMRMLScalarVolumeNode,
     vtkMRMLMarkupsFiducialNode,
 )
-from slicer.parameterNodeWrapper import (
-    parameterNodeSerializer,
-    parameterPack,
-    Serializer,
-    ValidatedSerializer,
-    validators,
-)
+from slicer.parameterNodeWrapper import parameterPack
 from slicer.parameterNodeWrapper.serializers import createSerializerFromAnnotatedType
-import importlib
-import sys
 import logging
-import zlib
-import io
-import base64
-
-
+from OpenLIFULib.lazyimport import openlifu_lz, xarray_lz
+from OpenLIFULib.parameter_node_utils import (
+    SlicerOpenLIFUTransducerWrapper,
+    SlicerOpenLIFUPoint,
+    SlicerOpenLIFUXADataset,
+    SlicerOpenLIFUProtocol,
+)
 if TYPE_CHECKING:
     import openlifu # This import is deferred at runtime, but it is done here for IDE and static analysis purposes
     import xarray
 
-class BusyCursor:
-    """
-    Context manager for showing a busy cursor.  Ensures that cursor reverts to normal in
-    case of an exception.
-    """
-
-    def __enter__(self):
-        qt.QApplication.setOverrideCursor(qt.Qt.BusyCursor)
-
-    def __exit__(self, exception_type, exception_value, traceback):
-        qt.QApplication.restoreOverrideCursor()
-        return False
-
-def install_python_requirements() -> None:
-    """Install python requirements"""
-    requirements_path = Path(__file__).parent / 'Resources/python-requirements.txt'
-    with BusyCursor():
-        slicer.util.pip_install(['-r', requirements_path])
-
-def python_requirements_exist() -> bool:
-    """Check and return whether python requirements are installed."""
-    return importlib.util.find_spec('openlifu') is not None
-
-def check_and_install_python_requirements(prompt_if_found = False) -> None:
-    """Check whether python requirements are installed and prompt to install them if not.
-
-    Args:
-        prompt_if_found: If this is enabled then in the event that python requirements are found,
-            there is a further prompt asking whether to run the install anyway.
-    """
-    want_install = False
-    if not python_requirements_exist():
-        want_install = slicer.util.confirmYesNoDisplay(
-            text = "Some OpenLIFU python dependencies were not found. Install them now?",
-            windowTitle = "Install python dependencies?",
-        )
-    elif prompt_if_found:
-        want_install = slicer.util.confirmYesNoDisplay(
-            text = "All OpenLIFU python dependencies were found. Re-run the install command?",
-            windowTitle = "Reinstall python dependencies?",
-        )
-    if want_install:
-        install_python_requirements()
-        if python_requirements_exist():
-            slicer.util.infoDisplay(text="Python requirements installed.", windowTitle="Success")
-        else:
-            slicer.util.errorDisplay(
-                text="OpenLIFU python dependencies are still not found. The install may have failed.",
-                windowTitle="Python dependencies still not found"
-            )
-
-def import_openlifu_with_check() -> "openlifu":
-    """Import openlifu and return the module, checking that it is installed along the way."""
-    if "openlifu" not in sys.modules:
-        check_and_install_python_requirements(prompt_if_found=False)
-        with BusyCursor():
-            import openlifu
-    return sys.modules["openlifu"]
-openlifu_lz = import_openlifu_with_check # A handy alternative short name. Stands for "openlifu lazy import"
-
-def import_xarray_with_check() -> "xarray":
-    """Import xarray and return the module, checking that openlifu is installed along the way."""
-    if "openlifu" not in sys.modules:
-        check_and_install_python_requirements(prompt_if_found=False)
-        import xarray
-    return sys.modules["xarray"]
-xarray_lz = import_xarray_with_check # A handy alternative short name. Stands for "openlifu lazy import"
+__all__ = [
+    "openlifu_lz",
+    "xarray_lz",
+    "SlicerOpenLIFUPlan",
+    "SlicerOpenLIFUProtocol",
+    "SlicerOpenLIFUTransducer",
+    "SlicerOpenLIFUPoint",
+    "SlicerOpenLIFUXADataset",
+    "PlanFocus",
+    "display_errors",
+    "create_noneditable_QStandardItem",
+    "ensure_list",
+    "add_slicer_log_handler",
+    "get_xxx2ras_matrix",
+    "get_xx2mm_scale_factor",
+    "fiducial_to_openlifu_point_in_transducer_coords",
+    "make_volume_from_xarray_in_transducer_coords",
+    "make_xarray_in_transducer_coords_from_volume",
+]
 
 def display_errors(f):
     """Decorator to make functions forward their python exceptions along as slicer error displays"""
@@ -186,7 +131,7 @@ def get_xxx2ras_matrix(dims:Sequence[str]) -> NDArray[Any]:
     ]).transpose()
 
 def get_xx2mm_scale_factor(length_unit:str) -> float:
-    openlifu = import_openlifu_with_check()
+    openlifu = openlifu_lz()
     return openlifu.util.units.getsiscale(length_unit, 'distance') / openlifu.util.units.getsiscale('mm', 'distance')
 
 def linear_to_affine(matrix, translation=None):
@@ -292,378 +237,6 @@ def make_xarray_in_transducer_coords_from_volume(volume_node:vtkMRMLScalarVolume
         attrs={'vtkMRMLNodeID':volume_node.GetID(),}
     )
     return volume_resampled_dataarray
-
-# This very thin wrapper around openlifu.Protocol is needed to do our lazy importing of openlifu
-# while still providing type annotations that the parameter node wrapper can use.
-# If we tried to make openlifu.Protocol directly supported as a type by parameter nodes, we would
-# get errors from parameterNodeWrapper as it tries to use typing.get_type_hints. This fails because
-# get_type_hints tries to *evaluate* the type annotations like "openlifu.Protocol" possibly before
-# the user has installed openlifu, and possibly before the main window widgets exist that would allow
-# an install prompt to even show up.
-class SlicerOpenLIFUProtocol:
-    """Ultrathin wrapper of openlifu.Protocol. This exists so that protocols can have parameter node
-    support while we still do lazy-loading of openlifu."""
-    def __init__(self, protocol: "openlifu.Protocol"):
-        self.protocol = protocol
-
-# For the same reason we have a thin wrapper around openlifu.Transducer. But the name SlicerOpenLIFUTransducer
-# is reserved for the upcoming parameter pack.
-class SlicerOpenLIFUTransducerWrapper:
-    """Ultrathin wrapper of openlifu.Transducer. This exists so that transducers can have parameter node
-    support while we still do lazy-loading of openlifu."""
-    def __init__(self, transducer: "openlifu.Transducer"):
-        self.transducer = transducer
-
-# For the same reason we have a thin wrapper around openlifu.Point
-class SlicerOpenLIFUPoint:
-    """Ultrathin wrapper of openlifu.Point. This exists so that points can have parameter node
-    support while we still do lazy-loading of openlifu."""
-    def __init__(self, point: "openlifu.Point"):
-        self.point = point
-
-# For the same reason we have a thin wrapper around xarray.Dataset
-class SlicerOpenLIFUXADataset:
-    """Ultrathin wrapper of xarray.Dataset, so that it can have parameter node
-    support while we still do lazy-loading of xarray (a dependency that is installed alongside openlifu)."""
-    def __init__(self, dataset: "xarray.Dataset"):
-        self.dataset = dataset
-
-@parameterNodeSerializer
-class OpenLIFUProtocolSerializer(Serializer):
-    @staticmethod
-    def canSerialize(type_) -> bool:
-        """
-        Whether the serializer can serialize the given type if it is properly instantiated.
-        """
-        return type_ == SlicerOpenLIFUProtocol
-
-    @staticmethod
-    def create(type_):
-        """
-        Creates a new serializer object based on the given type. If this class does not support the given type,
-        None is returned.
-        """
-        if OpenLIFUProtocolSerializer.canSerialize(type_):
-            # Add custom validators as we need them to the list here. For now just IsInstance.
-            return ValidatedSerializer(OpenLIFUProtocolSerializer(), [validators.IsInstance(SlicerOpenLIFUProtocol)])
-        return None
-
-    def default(self):
-        """
-        The default value to use if another default is not specified.
-        """
-        return SlicerOpenLIFUProtocol(openlifu_lz().Protocol())
-
-    def isIn(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> bool:
-        """
-        Whether the parameterNode contains a parameter of the given name.
-        Note that most implementations can just use parameterNode.HasParameter(name).
-        """
-        return parameterNode.HasParameter(name)
-
-    def write(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str, value: SlicerOpenLIFUProtocol) -> None:
-        """
-        Writes the value to the parameterNode under the given name.
-        """
-        parameterNode.SetParameter(
-            name,
-            value.protocol.to_json(compact=True)
-        )
-
-    def read(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> SlicerOpenLIFUProtocol:
-        """
-        Reads and returns the value with the given name from the parameterNode.
-        """
-        json_string = parameterNode.GetParameter(name)
-        return SlicerOpenLIFUProtocol(openlifu_lz().Protocol.from_json(json_string))
-
-    def remove(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> None:
-        """
-        Removes the value of the given name from the parameterNode.
-        """
-        parameterNode.UnsetParameter(name)
-
-@parameterNodeSerializer
-class OpenLIFUTransducerSerializer(Serializer):
-    @staticmethod
-    def canSerialize(type_) -> bool:
-        """
-        Whether the serializer can serialize the given type if it is properly instantiated.
-        """
-        return type_ == SlicerOpenLIFUTransducerWrapper
-
-    @staticmethod
-    def create(type_):
-        """
-        Creates a new serializer object based on the given type. If this class does not support the given type,
-        None is returned.
-        """
-        if OpenLIFUTransducerSerializer.canSerialize(type_):
-            # Add custom validators as we need them to the list here. For now just IsInstance.
-            return ValidatedSerializer(OpenLIFUTransducerSerializer(), [validators.IsInstance(SlicerOpenLIFUTransducerWrapper)])
-        return None
-
-    def default(self):
-        """
-        The default value to use if another default is not specified.
-        """
-        return SlicerOpenLIFUTransducerWrapper(openlifu_lz().Transducer())
-
-    def isIn(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> bool:
-        """
-        Whether the parameterNode contains a parameter of the given name.
-        Note that most implementations can just use parameterNode.HasParameter(name).
-        """
-        return parameterNode.HasParameter(name)
-
-    def write(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str, value: SlicerOpenLIFUTransducerWrapper) -> None:
-        """
-        Writes the value to the parameterNode under the given name.
-        """
-        parameterNode.SetParameter(
-            name,
-            value.transducer.to_json(compact=True)
-        )
-
-    def read(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> SlicerOpenLIFUTransducerWrapper:
-        """
-        Reads and returns the value with the given name from the parameterNode.
-        """
-        json_string = parameterNode.GetParameter(name)
-        return SlicerOpenLIFUTransducerWrapper(openlifu_lz().Transducer.from_json(json_string))
-
-    def remove(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> None:
-        """
-        Removes the value of the given name from the parameterNode.
-        """
-        parameterNode.UnsetParameter(name)
-
-@parameterNodeSerializer
-class OpenLIFUPointSerializer(Serializer):
-    @staticmethod
-    def canSerialize(type_) -> bool:
-        """
-        Whether the serializer can serialize the given type if it is properly instantiated.
-        """
-        return type_ == SlicerOpenLIFUPoint
-
-    @staticmethod
-    def create(type_):
-        """
-        Creates a new serializer object based on the given type. If this class does not support the given type,
-        None is returned.
-        """
-        if OpenLIFUPointSerializer.canSerialize(type_):
-            # Add custom validators as we need them to the list here. For now just IsInstance.
-            return ValidatedSerializer(OpenLIFUPointSerializer(), [validators.IsInstance(SlicerOpenLIFUPoint)])
-        return None
-
-    def default(self):
-        """
-        The default value to use if another default is not specified.
-        """
-        return SlicerOpenLIFUPoint(openlifu_lz().Point())
-
-    def isIn(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> bool:
-        """
-        Whether the parameterNode contains a parameter of the given name.
-        Note that most implementations can just use parameterNode.HasParameter(name).
-        """
-        return parameterNode.HasParameter(name)
-
-    def write(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str, value: SlicerOpenLIFUPoint) -> None:
-        """
-        Writes the value to the parameterNode under the given name.
-        """
-        parameterNode.SetParameter(
-            name,
-            value.point.to_json(compact=True)
-        )
-
-    def read(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> SlicerOpenLIFUPoint:
-        """
-        Reads and returns the value with the given name from the parameterNode.
-        """
-        json_string = parameterNode.GetParameter(name)
-        return SlicerOpenLIFUPoint(openlifu_lz().Point.from_json(json_string))
-
-    def remove(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> None:
-        """
-        Removes the value of the given name from the parameterNode.
-        """
-        parameterNode.UnsetParameter(name)
-
-@parameterNodeSerializer
-class XarraydatasetSerializer(Serializer):
-    @staticmethod
-    def canSerialize(type_) -> bool:
-        """
-        Whether the serializer can serialize the given type if it is properly instantiated.
-        """
-        return type_ == SlicerOpenLIFUXADataset
-
-    @staticmethod
-    def create(type_):
-        """
-        Creates a new serializer object based on the given type. If this class does not support the given type,
-        None is returned.
-        """
-        if XarraydatasetSerializer.canSerialize(type_):
-            # Add custom validators as we need them to the list here. For now just IsInstance.
-            return ValidatedSerializer(XarraydatasetSerializer(), [validators.IsInstance(SlicerOpenLIFUXADataset)])
-        return None
-
-    def default(self):
-        """
-        The default value to use if another default is not specified.
-        """
-        return SlicerOpenLIFUXADataset(xarray_lz().Dataset())
-
-    def isIn(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> bool:
-        """
-        Whether the parameterNode contains a parameter of the given name.
-        Note that most implementations can just use parameterNode.HasParameter(name).
-        """
-        return parameterNode.HasParameter(name)
-
-    def write(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str, value: SlicerOpenLIFUXADataset) -> None:
-        """
-        Writes the value to the parameterNode under the given name.
-        """
-        ds = value.dataset
-        ds_serialized = base64.b64encode(ds.to_netcdf()).decode('utf-8')
-        parameterNode.SetParameter(
-            name,
-            ds_serialized,
-        )
-
-    def read(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> SlicerOpenLIFUXADataset:
-        """
-        Reads and returns the value with the given name from the parameterNode.
-        """
-        ds_serialized = parameterNode.GetParameter(name)
-        ds_deserialized = xarray_lz().open_dataset(base64.b64decode(ds_serialized.encode('utf-8')))
-        return SlicerOpenLIFUXADataset(ds_deserialized)
-
-    def remove(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> None:
-        """
-        Removes the value of the given name from the parameterNode.
-        """
-        parameterNode.UnsetParameter(name)
-
-@parameterNodeSerializer
-class NumpyArraySerializer(Serializer):
-    @staticmethod
-    def canSerialize(type_) -> bool:
-        """
-        Whether the serializer can serialize the given type if it is properly instantiated.
-        """
-        return type_ == np.ndarray
-
-    @staticmethod
-    def create(type_):
-        """
-        Creates a new serializer object based on the given type. If this class does not support the given type,
-        None is returned.
-        """
-        if NumpyArraySerializer.canSerialize(type_):
-            # Add custom validators as we need them to the list here. For now just IsInstance.
-            return ValidatedSerializer(NumpyArraySerializer(), [validators.IsInstance(np.ndarray)])
-        return None
-
-    def default(self):
-        """
-        The default value to use if another default is not specified.
-        """
-        return np.array([])
-
-    def isIn(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> bool:
-        """
-        Whether the parameterNode contains a parameter of the given name.
-        Note that most implementations can just use parameterNode.HasParameter(name).
-        """
-        return parameterNode.HasParameter(name)
-
-    def write(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str, value: np.ndarray) -> None:
-        """
-        Writes the value to the parameterNode under the given name.
-        """
-        buffer = io.BytesIO()
-        np.save(buffer, value)
-        array_serialized = base64.b64encode(zlib.compress(buffer.getvalue())).decode('utf-8')
-        parameterNode.SetParameter(
-            name,
-            array_serialized,
-        )
-
-    def read(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> np.ndarray:
-        """
-        Reads and returns the value with the given name from the parameterNode.
-        """
-        array_serialized = parameterNode.GetParameter(name)
-        array_deserialized = np.load(io.BytesIO(zlib.decompress(base64.b64decode(array_serialized.encode('utf-8')))))
-        return array_deserialized
-
-    def remove(self, parameterNode: slicer.vtkMRMLScriptedModuleNode, name: str) -> None:
-        """
-        Removes the value of the given name from the parameterNode.
-        """
-        parameterNode.UnsetParameter(name)
-
-@parameterNodeSerializer
-class NamedTupleSerializer(Serializer):
-    """Serializer for NamedTuple largely copied from slicer.util.parameterNodeWrapper.serializers.TupleSerializer.
-    """
-    @staticmethod
-    def canSerialize(type_) -> bool:
-        return issubclass(type_, tuple) and hasattr(type_, '_fields') and callable(type_) and isinstance(type_,type)
-
-    @staticmethod
-    def create(type_):
-        if NamedTupleSerializer.canSerialize(type_):
-            args = tuple(type_.__annotations__[f] for f in type_._fields)
-            if len(args) == 0:
-                raise Exception("Unsure how to handle a typed tuple with no discernible type")
-            serializers = [createSerializerFromAnnotatedType(arg) for arg in args]
-            return NamedTupleSerializer(serializers, type_)
-        return None
-
-    def __init__(self, serializers, cls):
-        self._len = len(serializers)
-        self._serializers = serializers
-        self._fields = cls._fields
-        self._cls = cls
-
-    def default(self):
-        return self._cls(**{f:s.default() for f,s in zip(self._fields,self._serializers)})
-
-    @staticmethod
-    def _paramName(name, field):
-        return f"{name}.{field}"
-
-    def isIn(self, parameterNode, name: str) -> bool:
-        return self._serializers[0].isIn(parameterNode, self._paramName(name, self._fields[0]))
-
-    def write(self, parameterNode, name: str, value) -> None:
-        with slicer.util.NodeModify(parameterNode):
-            for field, serializer in zip(self._fields, self._serializers):
-                serializer.write(parameterNode, self._paramName(name, field), getattr(value,field))
-
-    def read(self, parameterNode, name: str):
-        return self._cls(
-            **{
-                field : serializer.read(parameterNode, self._paramName(name, field))
-                for field, serializer in zip(self._fields, self._serializers)
-            }
-        )
-
-    def remove(self, parameterNode, name: str) -> None:
-        with slicer.util.NodeModify(parameterNode):
-            for field, serializer in zip(self._fields,self._serializers):
-                serializer.remove(parameterNode, self._paramName(name, field))
-
-    def supportsCaching(self):
-        return all([s.supportsCaching() for s in self._serializers])
 
 @parameterPack
 class SlicerOpenLIFUTransducer:
