+++ OpenLIFULib/OpenLIFULib/busycursor.py
@@ -1,0 +1,16 @@
+"""Busy cursor context manager"""
+
+import qt
+
+class BusyCursor:
+    """
+    Context manager for showing a busy cursor.  Ensures that cursor reverts to normal in
+    case of an exception.
+    """
+
+    def __enter__(self):
+        qt.QApplication.setOverrideCursor(qt.Qt.BusyCursor)
+
+    def __exit__(self, exception_type, exception_value, traceback):
+        qt.QApplication.restoreOverrideCursor()
+        return False
